Fix issue with removing the cwd
