+++ tests/conftest.py
@@ -1,3 +1,6 @@
+import os
+from tempfile import TemporaryDirectory
+
 import pytest
 
 from openapi_spec_tools.base_gen.files import set_copyright
@@ -8,3 +11,12 @@ def copyright_fixture():
     set_copyright()  # set to default
     yield
     set_copyright() # reset to default
+
+
+@pytest.fixture
+def temp_working_dir():
+    orig = os.getcwd()
+    with TemporaryDirectory() as temp_dir:
+        os.chdir(temp_dir)
+        yield temp_dir
+        os.chdir(orig)
